# perf(init): Prefer set -x over export for fish users